Added Responsiveness to OCR UI
Upon selecting a file, the modal will display a loading screen while it uploads the file. It will then close automatically. When uploading a file, the text input box is replaced by an image of a file and the name of the uploaded file. This file can be dismissed to bring back the text input. Uploading a file will activate the search button